Extract update and lockfile functionality to separate files
